refactor: Create handler for boss bars
