+++ common/src/main/java/com/wynntils/handlers/bossbar/BossBarProgress.java
@@ -1,0 +1,7 @@
+/*
+ * Copyright © Wynntils 2022.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.handlers.bossbar;
+
+public record BossBarProgress(int current, int max, float progress) {}
