+++ common/src/main/java/com/wynntils/wynn/model/bossbar/BarType.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright © Wynntils 2022.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.wynn.model.bossbar;
+
+public enum BarType {
+    BLOODPOOL,
+    MANABANK,
+    AWAKENED,
+    FOCUS,
+    CORRUPTED
+}
